refactor(ResultsDtoFormatter): ResultDtosFormatter의 인자를 ResultsDto로 변경 및 클래스 이름 변경
